Update gen test to use nanoTime instead of cpu time
CPU time seems useless on my machine, and off by at least
an order of magnitude.

By timing each executable using nanoTime, should get a close
enough answer.

Generally, the generation test is useless due to how the Vanilla
light engine functions. As the number of loaded chunks increase,
so does the size of the maps visibleSectionData/updatingSectionData
in LayerLightSectionStorage. This results in
LayerLightSectionStorage#swapSectionMap dominating (>70% of the
time on Vanilla) the test result, which isn't what we're interested
in testing.